+++ CardanoSharp.Wallet.Test/NativeScriptTests.cs
@@ -1,8 +1,7 @@
 ﻿using CardanoSharp.Wallet.Extensions;
 using CardanoSharp.Wallet.Extensions.Models;
-using CardanoSharp.Wallet.Models.Transactions;
-using CardanoSharp.Wallet.Models.Transactions.Scripts;
-using CardanoSharp.Wallet.Models.Transactions.TransactionWitness.Scripts;
+using CardanoSharp.Wallet.Models.Transactions.TransactionWitness;
+using CardanoSharp.Wallet.Models.Transactions.TransactionWitness.NativeScripts;
 using Xunit;
 
 namespace CardanoSharp.Wallet.Test;
