+++ CardanoSharp.Wallet/Enums/RedeemerTag.cs
@@ -1,0 +1,10 @@
+﻿namespace CardanoSharp.Wallet.Enums
+{
+    public enum RedeemerTag
+    {
+        Spend = 0,
+        Mint = 1,
+        Cert = 2, 
+        Reward = 3
+    }
+}
