+++ CardanoSharp.Wallet/Extensions/Models/ScriptInvalidAfterExtension.cs
@@ -1,4 +1,4 @@
-﻿using CardanoSharp.Wallet.Models.Transactions.TransactionWitness.Scripts;
+﻿using CardanoSharp.Wallet.Models.Transactions.TransactionWitness.NativeScripts;
 using PeterO.Cbor2;
 using System;
