+++ CardanoSharp.Wallet/Extensions/Models/ScriptPubKeyExtension.cs
@@ -1,4 +1,4 @@
-﻿using CardanoSharp.Wallet.Models.Transactions.Scripts;
+﻿using CardanoSharp.Wallet.Models.Transactions.TransactionWitness.NativeScripts;
 using PeterO.Cbor2;
 using System;
