+++ CardanoSharp.Wallet/Models/Transactions/TransactionWitness/PlutusScripts/ExUnits.cs
@@ -1,0 +1,8 @@
+﻿namespace CardanoSharp.Wallet.Models.Transactions.TransactionWitness.PlutusScripts
+{
+    public class ExUnits
+    {
+        public uint Mem { get; set; }
+        public uint Steps { get; set; }
+    }
+}
