+++ CardanoSharp.Wallet/Models/Transactions/TransactionWitness/PlutusScripts/Redeemer.cs
@@ -1,0 +1,12 @@
+﻿using CardanoSharp.Wallet.Enums;
+
+namespace CardanoSharp.Wallet.Models.Transactions.TransactionWitness.PlutusScripts
+{
+    public class Redeemer
+    {
+        public RedeemerTag Tag { get; set; }
+        public uint Index { get; set; }
+        public object PlutusData { get; set; }
+        public ExUnits ExUnits { get; set; }
+    }
+}
