+++ CardanoSharp.Wallet/Models/Transactions/TransactionWitness/TransactionWitnessSet.cs
@@ -1,17 +1,19 @@
 ﻿using System.Collections.Generic;
+using CardanoSharp.Wallet.Models.Transactions.TransactionWitness.PlutusScripts;
 
-namespace CardanoSharp.Wallet.Models.Transactions
+namespace CardanoSharp.Wallet.Models.Transactions.TransactionWitness
 {
     //MultiSig: https://docs.cardano.org/projects/cardano-node/en/latest/reference/simple-scripts.html#type-all
 
-    //  transaction_witness_set =
-    //{ ? 0: [* vkeywitness ]
-    //, ? 1: [* native_script ]
-    //, ? 2: [* bootstrap_witness ]
-    //; In the future, new kinds of witnesses can be added like this:
-    //; , ? 4: [* foo_script]
-    //; , ? 5: [* plutus_script]
-    //}
+    // transaction_witness_set =
+    // { ? 0: [* vkeywitness ]
+    //         , ? 1: [* native_script ]
+    //         , ? 2: [* bootstrap_witness ]
+    //         , ? 3: [* plutus_v1_script ]
+    //         , ? 4: [* plutus_data ]
+    //         , ? 5: [* redeemer ]
+    //         , ? 6: [* plutus_v2_script ] ; New
+    // }
 
     //This can be updated for future witness changes based upon CDDL changes
     public partial class TransactionWitnessSet
@@ -24,7 +26,10 @@ namespace CardanoSharp.Wallet.Models.Transactions
         public ICollection<VKeyWitness> VKeyWitnesses { get; set; }
         public ICollection<NativeScript> NativeScripts { get; set; }
         public ICollection<BootStrapWitness> BootStrapWitnesses { get; set; }
-
-
+        public ICollection<byte[]> PlutusScriptV1 { get; set; }
+        public ICollection<object> PlutusData { get; set; }
+        public ICollection<Redeemer> Redeemer { get; set; }
+        public ICollection<byte[]> PlutusScriptV2 { get; set; }
+        
     }
 }
