+++ CardanoSharp.Wallet/TransactionBuilding/NativeScriptBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using CardanoSharp.Wallet.Enums;
 using CardanoSharp.Wallet.Models.Transactions;
-using CardanoSharp.Wallet.Models.Transactions.Scripts;
-using CardanoSharp.Wallet.Models.Transactions.TransactionWitness.Scripts;
+using CardanoSharp.Wallet.Models.Transactions.TransactionWitness;
+using CardanoSharp.Wallet.Models.Transactions.TransactionWitness.NativeScripts;
 using System.Collections.Generic;
 
 namespace CardanoSharp.Wallet.TransactionBuilding
